Changing the behavior surrounding see pointers and nested versioning.
Previously if an annotation that was being loaded through an `@api-see`
pointer had any versioning of its own set, that would be overridden in
favor of the versioning applied to the `@api-see` call.

Now, however, those two versions are now combined into a new constraint.

So, for example, `external_urls` has `>=1.1`, but loads a method that
has `external_urls.tickets` at `<1.1.3`. Previously `external_urls.tickets`
would become `>1.1.`, so with this change `external_urls.tickets` is
now versioned at `>1.1 <1.1.3`.